+++ version_c/Project1/utils.cpp
@@ -1,16 +1,16 @@
 #include "utils.h"
 
-float dist(float* p1, float* p2, Config* config)
+double dist(double* p1, double* p2, Config* config)
 {
-    float dist = 0.0;
+    double dist = 0.0;
     for (int i = 0; i < config->NUM_DIM; i++)
     {
-        dist += (p1[i] - p2[i]) * (p1[i] - p2[i]);
+        dist += (p1[i] - p2[i])* (p1[i] - p2[i]);
     }
     return dist;
 }
 
-bool same_points(float* p1, float* p2, Config* config)
+bool same_points(double* p1, double* p2, Config* config)
 {
     bool issame = true;
     for (int i = 0; i < config->NUM_DIM; i++) {
@@ -22,7 +22,7 @@ bool same_points(float* p1, float* p2, Config* config)
     return issame;
 }
 
-float prandom(float min, float max)
+double prandom(double min, double max)
 {
-    return min + ((float)rand() / (float)(RAND_MAX)) * (max - min);
+    return min + ((double)rand() / (double)(RAND_MAX)) * (max - min);
 }
